Bug fix: wrong 'git mv' commands, sometimes
Fix for the bug that sometimes causes the wrong 'git mv' commands to be
generated (apparently due to being too liberal in the original search for
matches for the old-module-name).  First reported in the email to
bug-Module-Rename@rt.cpan.org with the subject:

Thanks, and found new bug [Re: [rt.cpan.org #87172] Request to handle 'git mv' in Module::Rename]

Further description of the bug is provided in follow-up emails to
bug-Module-Rename@rt.cpan.org in the same thread.